+++ memtable/write_buffer_manager.cc
@@ -745,7 +745,6 @@ void WriteBufferManager::InitiateFlushesThread() {
           // the callback (which may take a long time).
           was_flush_initiated = initiator.cb(kMinFlushSizes[iter]);
         }
-
         if (!was_flush_initiated) {
           // No flush was initiated => undo the counters update
           assert(num_running_flushes_ > 0U);
@@ -809,9 +808,9 @@ void WriteBufferManager::FlushEnded(bool /* wbm_initiated */) {
   // the WBM will not be aware of the number of running flushes at the time
   // it is enabled. The counter will become valid once all of the flushes
   // that were running when it was enabled will have completed.
-  if (num_running_flushes_ > 0U) {
-    --num_running_flushes_;
-  }
+  assert(num_running_flushes_ > 0U);
+  --num_running_flushes_;
+
   size_t curr_memory_used = memory_usage();
   RecalcFlushInitiationSize();
   ReevaluateNeedForMoreFlushesLockHeld(curr_memory_used);